♻️ refactor: 커스텀 훅으로 api호출 변경
