+++ services/quiz/src/pages/StudyDetailPage/models/quizList.ts
@@ -1,0 +1,11 @@
+import { User } from '../models/user';
+
+export interface QuizItem {
+    quizId: number;
+    weekNumber: number;
+    title: string;
+    openDate: Date;
+    limitTime: string;
+    creator: User;
+    absentee: User[];
+  }
