+++ services/quiz/src/pages/StudyDetailPage/models/rankingList.ts
@@ -1,0 +1,9 @@
+import { User } from "./user";
+
+type Ranking = number;
+
+export interface RankingItem {
+  ranking?: Ranking;
+  toks: number;
+  user: User;
+}
