+++ services/quiz/src/pages/StudyDetailPage/models/studyInfo.ts
@@ -1,0 +1,9 @@
+import { User } from "./user";
+
+export interface StudyInfo {
+    studyId: number;
+    title: string;
+    description: string;
+    studyTags: string[];
+    members: User[];
+  }
