+++ services/quiz/src/pages/StudyDetailPage/models/user.ts
@@ -1,0 +1,6 @@
+export type User = {
+    image: string;
+    id: string;
+    userName: string;
+    size?: string;
+  };
