+++ services/quiz/src/pages/StudyDetailPage/remotes/quizList.ts
@@ -1,0 +1,6 @@
+import { quizList } from '../../../mock/db';
+import { QuizItem } from '../models/quizList';
+
+export const getQuizList = async () => {
+    return await new Promise<QuizItem[]>(resolve => setTimeout(() => resolve(quizList), 3000));
+  };
