+++ services/quiz/src/pages/StudyDetailPage/remotes/rankingList.ts
@@ -1,0 +1,6 @@
+import { rankingList } from "mock/db";
+import { RankingItem } from "../models/rankingList";
+
+export const getRankingList = async () => {
+    return await new Promise<RankingItem[]>(resolve => setTimeout(() => resolve(rankingList), 1000));
+  };
