+++ services/quiz/src/pages/StudyDetailPage/remotes/studyInfo.ts
@@ -1,0 +1,6 @@
+import { studyInfo } from "mock/db";
+import { StudyInfo } from "../models/studyInfo";
+
+export const getStudyInfo = async () => {
+    return await new Promise<StudyInfo>(resolve => setTimeout(() => resolve(studyInfo), 2000));
+  };
